+++ front-web/src/styles/CalanderContents.scss
@@ -1,0 +1,18 @@
+.calander__dayOfWeek__wrapper {
+  box-sizing: border-box;
+  display: flex;
+  width: 100%;
+  margin-top: 10px;
+}
+.calander__dayOfWeek__days {
+  flex: 1;
+  text-align: right;
+  padding-right: 10px;
+  padding-bottom: 10px;
+}
+.calander__dayOfWeek__days:nth-child(1) {
+  color: red;
+}
+.calander__dayOfWeek__days:nth-child(7) {
+  color: blue;
+}
