+++ front-web/src/styles/Todo.scss
@@ -2,16 +2,6 @@
   display: none;
 }
 
-.show {
-  display: 'block';
-}
-.hide {
-  display: none;
-}
-.done {
-  color: blue;
-  text-decoration: line-through;
-}
 .todo_done {
   opacity: 0.3;
   text-decoration: line-through;
@@ -52,31 +42,6 @@
     flex: 8;
     overflow: scroll;
     overflow-x: hidden;
-  }
-
-  .todos__contents {
-    position: relative;
-    margin-bottom: 16px;
-    padding-top: 16px;
-    padding-bottom: 10px;
-    border-image-repeat: stretch;
-
-    .todos__delBtn {
-      // float: right;
-      position: absolute;
-      top: 0;
-      right: 5px;
-      font-weight: 700;
-    }
-    .todos__delBtn:hover {
-      color: red;
-    }
-    .todo__content {
-      word-break: break-all;
-    }
-  }
-  .todos__contents:hover {
-    background-color: rgba($color: #000000, $alpha: 0.1);
   }
 
   .togle_todo_btn {
