+++ front-web/src/styles/_global.scss
@@ -1,0 +1,22 @@
+* {
+  box-sizing: border-box;
+}
+.show {
+  display: block !important;
+}
+.hide {
+  display: hide !important;
+}
+
+.done {
+  text-decoration: line-through !important;
+}
+
+.grabbing {
+  background-color: pink !important;
+  cursor: move;
+}
+
+.unhighlight {
+  opacity: 0.2 !important;
+}
